+++ pkg/networking/transport/onramp_i2p/transport_test.go
@@ -1,4 +1,4 @@
-package onramp_i2p
+package onrampi2p
 
 import (
 	"context"