+++ pkg/networking/transport/onramp_tor/transport_test.go
@@ -1,4 +1,4 @@
-package onramp_tor
+package onramptor
 
 import (
 	"context"